Revert "move most root level code to scripts"
