+++ interpolate_series.py
@@ -2,9 +2,9 @@
 import argparse
 from typing import Callable
 from tqdm import tqdm
-from scripts.interpolate_engine import InterpolateEngine
-from scripts.interpolate import Interpolate
-from scripts.deep_interpolate import DeepInterpolate
+from interpolate_engine import InterpolateEngine
+from interpolate import Interpolate
+from deep_interpolate import DeepInterpolate
 from webui_utils.simple_log import SimpleLog
 from webui_utils.file_utils import create_directory, get_files
