+++ interpolation_target.py
@@ -6,8 +6,8 @@ from typing import Callable
 import re
 import cv2
 from tqdm import tqdm
-from scripts.interpolate_engine import InterpolateEngine
-from scripts.interpolate import Interpolate
+from interpolate_engine import InterpolateEngine
+from interpolate import Interpolate
 from webui_utils.simple_log import SimpleLog
 from webui_utils.simple_utils import float_range_in_range, sortable_float_index
 from webui_utils.file_utils import create_directory
