+++ restore_frames.py
@@ -3,9 +3,9 @@ import sys
 import argparse
 from typing import Callable
 from tqdm import tqdm
-from scripts.interpolate_engine import InterpolateEngine
-from scripts.interpolate import Interpolate
-from scripts.interpolation_target import TargetInterpolate
+from interpolate_engine import InterpolateEngine
+from interpolate import Interpolate
+from interpolation_target import TargetInterpolate
 from webui_utils.simple_log import SimpleLog
 from webui_utils.file_utils import create_directory
 from webui_utils.simple_utils import restored_frame_searches
