+++ tabs/future_ui.py
@@ -1,11 +1,11 @@
 """Future tabs UI and event handlers"""
 from typing import Callable
 import gradio as gr
-from scripts.interpolate_engine import InterpolateEngine
+from interpolate_engine import InterpolateEngine
 from webui_utils.simple_config import SimpleConfig
 from webui_utils.simple_icons import SimpleIcons
 
-from scripts.webui_tips import WebuiTips
+from webui_tips import WebuiTips
 
 class Future():
     """Encapsulates UI elements and events for the Future tabs"""
