+++ tabs/video_inflation_ui.py
@@ -6,11 +6,11 @@ from webui_utils.simple_icons import SimpleIcons
 from webui_utils.file_utils import create_directory, get_files
 from webui_utils.auto_increment import AutoIncrementDirectory
 from webui_utils.ui_utils import update_splits_info
-from scripts.webui_tips import WebuiTips
-from scripts.interpolate_engine import InterpolateEngine
-from scripts.interpolate import Interpolate
-from scripts.deep_interpolate import DeepInterpolate
-from scripts.interpolate_series import InterpolateSeries
+from webui_tips import WebuiTips
+from interpolate_engine import InterpolateEngine
+from interpolate import Interpolate
+from deep_interpolate import DeepInterpolate
+from interpolate_series import InterpolateSeries
 
 class VideoInflation():
     """Encapsulates UI elements and events for the Video Inflation feature"""
